% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normvols.r
\name{normbytcv}
\alias{normbytcv}
\title{Normalize ROI Volumes by Total Brain Volume}
\usage{
normbytcv(
  data,
  TBV_var,
  ROI_vars,
  subject_id_vars,
  proportion = 0.66,
  powers_set = c(-2, -1.5, -1, -0.5, "log", 0.5, 1, 1.5, 2, 3),
  num_cores = 1,
  plot_models = FALSE,
  save_results = FALSE,
  results_name = "normbytcv_results"
)
}
\arguments{
\item{data}{A data frame containing the ROI and TBV data.}

\item{TBV_var}{The name of the total brain volume variable as a string.}

\item{ROI_vars}{A vector of ROI variable names to normalize.}

\item{subject_id_vars}{A vector of identification variable names (e.g., subject ID, session ID).}

\item{proportion}{Define threshold for model bagging. The proportion (probability 0-1) that a model fits as well as the best model (default is 0.66).}

\item{powers_set}{A vector of powers for fractional polynomial transformations.
Default is c(-2, -1.5, -1, -0.5, "log", 0.5, 1, 1.5, 2, 3).}

\item{num_cores}{Number of cores to use for parallel processing (default 1).}

\item{plot_models}{Logical, if TRUE, plots the selected model functions for each ROI variable.}

\item{save_results}{Logical, if TRUE, saves the results object to an RDS file.}

\item{results_name}{A string specifying the name of the results object to save (without file extension).}
}
\value{
A list containing:
\describe{
  \item{adjusted\_volumes}{A data frame with the adjusted ROI volumes.}
  \item{results}{A list containing the models and adjusted ROI values for each ROI variable.}
}
}
\description{
This function normalizes ROI volumes by total brain volume using bagged fractional polynomial models.
Adjusts volumetric regions of interest for their association with overall brain volume using fractional polynomials.
}
\examples{
\dontrun{
norm\_results <- normbytcv(data, "TBV", c("ROI1", "ROI2"), c("ID"), 0.66, plot_models = TRUE)
}

}
